// src/character/models/character-gender.ts
/**
 *
 * @export
 * @interface CharacterGender
 */
export interface CharacterGender {
  /**
   *
   * @type {string}
   * @memberof CharacterGender
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof CharacterGender
   */
  name: string;
  /**
   *
   * @type {boolean}
   * @memberof CharacterGender
   */
  enabled: boolean;
  /**
   *
   * @type {number}
   * @memberof CharacterGender
   */
  order: number;
}