// src/character/models/character-nationality.ts
/**
 *
 * @export
 * @interface CharacterNationality
 */
export interface CharacterNationality {
  /**
   *
   * @type {string}
   * @memberof CharacterNationality
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof CharacterNationality
   */
  name: string;
  /**
   *
   * @type {boolean}
   * @memberof CharacterNationality
   */
  enabled: boolean;
  /**
   *
   * @type {number}
   * @memberof CharacterNationality
   */
  order: number;
}